+++ docker-compose.yml
@@ -29,7 +29,6 @@ services:
       - 4001:4001 # ACMEv2
       - 4002:4002 # OCSP
       - 4003:4003 # OCSP
-      - 4430:4430 # ACME via HTTPS
       - 4431:4431 # ACMEv2 via HTTPS
       - 8055:8055 # dns-test-srv updates
     depends_on:
